+++ betteruptime/betteruptime.py
@@ -1,10 +1,15 @@
+import asyncio
 import logging
+from typing import Optional
 
 import pandas
+import sentry_sdk
+import vexcogutils
 from redbot.core import Config, commands
 from redbot.core.bot import Red
 from redbot.core.utils.chat_formatting import pagify
 from vexcogutils import format_help, format_info
+from vexcogutils.meta import out_of_date_check
 
 from betteruptime.commands import BUCommands
 
@@ -29,7 +34,7 @@ class BetterUptime(commands.Cog, BUCommands, BULoop, Utils, metaclass=CompositeM
     data to become available.
     """
 
-    __version__ = "2.0.4"
+    __version__ = "2.0.5"
     __author__ = "Vexed#3211"
 
     def __init__(self, bot: Red) -> None:
@@ -61,6 +66,50 @@ class BetterUptime(commands.Cog, BUCommands, BULoop, Utils, metaclass=CompositeM
         except Exception:
             pass
 
+        asyncio.create_task(self.async_init())
+
+        # =========================================================================================
+        # NOTE: IF YOU ARE EDITING MY COGS, PLEASE ENSURE SENTRY IS DISBALED BY FOLLOWING THE INFO
+        # IN async_init(...) BELOW (SENTRY IS WHAT'S USED FOR TELEMETRY + ERROR REPORTING)
+        self.sentry_hub: Optional[sentry_sdk.Hub] = None
+        # =========================================================================================
+
+    async def async_init(self):
+        await out_of_date_check("betteruptime", self.__version__)
+
+        # =========================================================================================
+        # TO DISABLE SENTRY FOR THIS COG (EG IF YOU ARE EDITING THIS COG) EITHER DISABLE SENTRY
+        # WITH THE `[p]vextelemetry` COMMAND, OR UNCOMMENT THE LINE BELOW, OR REMOVE IT COMPLETELY:
+        # return
+
+        while vexcogutils.sentryhelper.ready is False:
+            await asyncio.sleep(0.1)
+
+        await vexcogutils.sentryhelper.maybe_send_owners("betteruptime")
+
+        if vexcogutils.sentryhelper.sentry_enabled is False:
+            _log.debug("Sentry detected as disabled.")
+            return
+
+        _log.debug("Sentry detected as enabled.")
+        self.sentry_hub = await vexcogutils.sentryhelper.get_sentry_hub(
+            "betteruptime", self.__version__
+        )
+        # =========================================================================================
+
+    async def cog_command_error(self, ctx: commands.Context, error: commands.CommandError):
+        await self.bot.on_command_error(ctx, error, unhandled_by_cog=True)  # type:ignore
+
+        if self.sentry_hub is None:  # sentry disabled
+            return
+
+        with self.sentry_hub:
+            sentry_sdk.add_breadcrumb(
+                category="command", message="Command used was " + ctx.command.qualified_name
+            )
+            sentry_sdk.capture_exception(error.original)  # type:ignore
+            _log.debug("Above exception successfully reported to Sentry")
+
     def format_help_for_context(self, ctx: commands.Context) -> str:
         """Thanks Sinbad."""
         return format_help(self, ctx)
@@ -83,12 +132,14 @@ class BetterUptime(commands.Cog, BUCommands, BULoop, Utils, metaclass=CompositeM
                 pass
             self.bot.add_command(old_uptime)
 
+        if self.sentry_hub:
+            self.sentry_hub.end_session()
+            self.sentry_hub.client.close()  # type:ignore
+
         try:
             self.bot.remove_dev_env_value("bu")
         except Exception:
             pass
-
-    # =============================================================================================
 
     @commands.command(hidden=True)
     async def betteruptimeinfo(self, ctx: commands.Context):