+++ timechannel/abc.py
@@ -1,9 +1,11 @@
 import asyncio
 from abc import ABC, ABCMeta
+from typing import Optional
 
 from redbot.core.bot import Red
 from redbot.core.commands import CogMeta
 from redbot.core.config import Config
+from sentry_sdk.hub import Hub
 from vexcogutils.loop import VexLoop
 
 
@@ -22,3 +24,5 @@ class MixinMeta(ABC):
 
     loop_meta: VexLoop
     loop: asyncio.Task
+
+    sentry_hub: Optional[Hub]