+++ wol/wol.py
@@ -1,15 +1,18 @@
-from typing import Dict
+import asyncio
+import logging
+from typing import Dict, Optional
 
+import sentry_sdk
 import tabulate
+import vexcogutils
 from redbot.core import commands
 from redbot.core.bot import Red
 from redbot.core.config import Config, Group
 from redbot.core.utils.chat_formatting import box
-from vexcogutils import format_help, format_info
+from vexcogutils import format_help, format_info, out_of_date_check
 from wakeonlan import send_magic_packet
 
-# NOTE: WHILST SEND_MAGIC_PACKET IS NOT ASYNC, IT DOES NOT WAIT FOR A RESPONSE SO I SEE LITTLE
-# REASON IN USING AN EXECUTOR. I HAVE TESTED IT AND IT TAKES >0.00 SECONDS
+log = logging.getLogger("red.vex.wol")
 
 
 def humanize_mac(mac: str):
@@ -28,7 +31,7 @@ class WOL(commands.Cog):
     `[p]wol main_pc`
     """
 
-    __version__ = "1.0.0"
+    __version__ = "1.0.5"
     __author__ = "Vexed#3211"
 
     def __init__(self, bot: Red) -> None:
@@ -37,6 +40,53 @@ class WOL(commands.Cog):
         self.config: Config = Config.get_conf(self, 418078199982063626, force_registration=True)
         self.config.register_global(version=1)
         self.config.register_global(addresses={})
+
+        asyncio.create_task(self.async_init())
+
+        # =========================================================================================
+        # NOTE: IF YOU ARE EDITING MY COGS, PLEASE ENSURE SENTRY IS DISBALED BY FOLLOWING THE INFO
+        # IN async_init(...) BELOW (SENTRY IS WHAT'S USED FOR TELEMETRY + ERROR REPORTING)
+        self.sentry_hub: Optional[sentry_sdk.Hub] = None
+        # =========================================================================================
+
+    async def async_init(self):
+        await out_of_date_check("wol", self.__version__)
+
+        # =========================================================================================
+        # TO DISABLE SENTRY FOR THIS COG (EG IF YOU ARE EDITING THIS COG) EITHER DISABLE SENTRY
+        # WITH THE `[p]vextelemetry` COMMAND, OR UNCOMMENT THE LINE BELOW, OR REMOVE IT COMPLETELY:
+        # return
+
+        while vexcogutils.sentryhelper.ready is False:
+            await asyncio.sleep(0.1)
+
+        await vexcogutils.sentryhelper.maybe_send_owners("wol")
+
+        if vexcogutils.sentryhelper.sentry_enabled is False:
+            log.debug("Sentry detected as disabled.")
+            return
+
+        log.debug("Sentry detected as enabled.")
+        self.sentry_hub = await vexcogutils.sentryhelper.get_sentry_hub("wol", self.__version__)
+        # =========================================================================================
+
+    async def cog_command_error(self, ctx: commands.Context, error: commands.CommandError):
+        await self.bot.on_command_error(ctx, error, unhandled_by_cog=True)  # type:ignore
+
+        if self.sentry_hub is None:  # sentry disabled
+            return
+
+        with self.sentry_hub:
+            sentry_sdk.add_breadcrumb(
+                category="command", message="Command used was " + ctx.command.qualified_name
+            )
+            sentry_sdk.capture_exception(error.original)  # type:ignore
+            log.debug("Above exception successfully reported to Sentry")
+
+    def cog_unload(self):
+        if self.sentry_hub:
+            self.sentry_hub.end_session()
+            self.sentry_hub.client.close()
 
     def format_help_for_context(self, ctx: commands.Context) -> str:
         """Thanks Sinbad."""
@@ -136,14 +186,11 @@ class WOL(commands.Cog):
 
         This will send your MAC addresses to the current channel.
         """
-        data = []
         conf: Dict[str, str] = await self.config.addresses()
         if not conf:
             return await ctx.send("You haven't added any machines yet.")
 
-        for name, mac in conf.items():
-            data.append([f"[{name}]", humanize_mac(mac)])
-
+        data = [[f"[{name}]", humanize_mac(mac)] for name, mac in conf.items()]
         table = tabulate.tabulate(data, headers=["Name", "MAC"])
 
         # god i hope no-one hits 2k chars